requirements.txt
fastapi[all]
motor
python-dotenv
uvicorn
pytest
pytest-asyncio
httpx
